fix(#2175 V2-S3b): move boxStrictEqOperand into value-tags.ts (#2104 box-site gate)
The #2104 AnyValue box-site gate requires all direct `funcMap.get("__any_box_*")`
emission to live in the sanctioned `value-tags.ts` (the boxToAny home) or
`any-helpers.ts` (the helper definitions). Relocate the strict-eq operand box-site
`boxStrictEqOperand` from coercion-engine.ts into value-tags.ts (exported) and
import it. Pure move — identical bytes; box-site + coercion-site gates now green.


✓
